feat(errors): rewrite API to use anyhow::Result instead of panicking
Switch to anyhow and thiserror crates instead of the failure crate for
big wins.

This should significantly improve error output on unexpected and
miscellaneous OS/IO errors. It also lays the groundwork for a more
official programmatic API.

This is technically breaking at the API level, but srisum has never
documented the API or treated it as official, so I'm going to pretend
this is a small feature instead. In the future, the API will actually be
official and have a nicer interface.